- Added support for ErrorBoundaries - componentDidCatch and getDerivedStateFromError - Move updateComponentNode as separate file as it was becoming bigger. - Return oldState, if newState is a false value in mergeState method.
