+++ src/updateNode.js
@@ -1,6 +1,5 @@
 import TemplateNode from './TemplateNode';
-import functionalComponentInstance from './functionalComponentInstance';
-import { PureComponent } from './Component';
+import updateComponentNode from './updateComponentNode';
 
 import {
   isNonZeroFalsy,
@@ -12,7 +11,6 @@ import {
   lastItem,
   toArray,
   removeNodes,
-  mergeState,
 } from './utils';
 
 import getTagNode from './TagNode';
@@ -85,89 +83,6 @@ function updateArrayNodes (part, nodes, oldNodes = []) {
   if (lastChild) deleteNodesBetween(parentNode, lastChild, nextSibling);
 
   return lastChild;
-}
-
-/**
- * Update component node
- */
-function updateComponentNode (part, node, oldNode, forceRender) {
-  const {
-    type: Component,
-    props,
-    __$isBrahmosClassComponent$__: isClassComponent,
-    __$isBrahmosFunctionalComponent$__: isFunctionalComponent,
-  } = node;
-
-  let firstRender = false;
-  let shouldUpdate = true;
-
-  /** If Component instance is not present on node create a new instance */
-  let { componentInstance } = node;
-
-  if (!componentInstance) {
-    // create an instance of the component
-    componentInstance = isFunctionalComponent
-      ? functionalComponentInstance(Component)
-      : new Component(props);
-
-    /**
-     * store the part and node information on the component instance,
-     * so every component have the createElement instance of self and
-     * the information of where it has to render
-     */
-    componentInstance.__part = part;
-    componentInstance.__componentNode = node;
-
-    // keep the reference of instance to the node.
-    node.componentInstance = componentInstance;
-
-    firstRender = true;
-  }
-
-  // call the life cycle methods for class component, which comes before rendering
-  if (isClassComponent) {
-    const { __unCommittedState, shouldComponentUpdate } = componentInstance;
-    let state = __unCommittedState || componentInstance.state;
-    // call getDerivedStateFromProps hook with the unCommitted state
-    const { getDerivedStateFromProps } = Component;
-    if (getDerivedStateFromProps) {
-      state = mergeState(
-        state,
-        getDerivedStateFromProps(props, state)
-      );
-    }
-
-    /**
-     * check if component is instance of PureComponent, if yes then,
-     * do shallow check for props and states
-     */
-
-    if (componentInstance instanceof PureComponent) {
-      shouldUpdate = state !== componentInstance.state || props !== componentInstance.props;
-    }
-
-    /**
-     * check if component should update or not. If PureComponent shallow check has already
-     * marked component to not update then we don't have to call shouldComponentUpdate
-     * Also we shouldn't call shouldComponentUpdate on first render
-     */
-    if (shouldComponentUpdate && shouldUpdate && !firstRender) {
-      shouldUpdate = shouldComponentUpdate(props);
-    }
-
-    // set the new state and props and reset uncommitted state
-    componentInstance.state = state;
-    componentInstance.props = props;
-    componentInstance.__unCommittedState = undefined;
-  }
-
-  // update a component update only if it can be updated based on shouldComponentUpdate
-  if (shouldUpdate) {
-    // render nodes
-    const renderNodes = componentInstance.__render(props);
-
-    return updateNode(part, renderNodes, null, forceRender);
-  }
 }
 
 /**
